refactor: extract tunnel wiring and discovery SSRF validation to siblings
