+++ src/synthorg/providers/discovery.py
@@ -12,14 +12,11 @@ URL probing (candidate URL probing for presets) lives in
 :mod:`synthorg.providers.probing`.
 """
 
-import asyncio
-import ipaddress
 import json
-import socket
 from collections.abc import Awaitable, Callable
 from contextvars import ContextVar
-from typing import Final, NamedTuple
-from urllib.parse import urlparse, urlunparse
+from typing import Final
+from urllib.parse import urlparse
 
 import httpx
 from pydantic import JsonValue
@@ -32,6 +29,10 @@ from synthorg.observability.events.provider import (
     PROVIDER_DISCOVERY_FAILED,
     PROVIDER_DISCOVERY_SSRF_BYPASSED,
     PROVIDER_MODELS_DISCOVERED,
+)
+from synthorg.providers._discovery_ssrf import (
+    build_pinned_url,
+    validate_discovery_url,
 )
 from synthorg.providers.capability_enrichment import (
     FetchContext,
@@ -58,155 +59,6 @@ _BYPASS_WARNED_ORIGINS: ContextVar[set[str] | None] = ContextVar(
 )
 
 _DISCOVERY_TIMEOUT_SECONDS: Final[float] = 10.0
-
-_ALLOWED_SCHEMES: Final[frozenset[str]] = frozenset({"http", "https"})
-
-# Private, loopback, link-local, and reserved networks.
-_BLOCKED_NETWORKS: Final[tuple[ipaddress.IPv4Network | ipaddress.IPv6Network, ...]] = (
-    ipaddress.IPv4Network("0.0.0.0/8"),
-    ipaddress.IPv4Network("10.0.0.0/8"),
-    ipaddress.IPv4Network("100.64.0.0/10"),
-    ipaddress.IPv4Network("127.0.0.0/8"),
-    ipaddress.IPv4Network("169.254.0.0/16"),
-    ipaddress.IPv4Network("172.16.0.0/12"),
-    ipaddress.IPv4Network("192.0.0.0/24"),
-    ipaddress.IPv4Network("192.0.2.0/24"),
-    ipaddress.IPv4Network("192.168.0.0/16"),
-    ipaddress.IPv6Network("::/128"),
-    ipaddress.IPv6Network("::1/128"),
-    ipaddress.IPv6Network("fc00::/7"),
-    ipaddress.IPv6Network("fe80::/10"),
-)
-
-
-class _SsrfCheckResult(NamedTuple):
-    """Result of SSRF URL validation.
-
-    Attributes:
-        error: Error message if the URL is unsafe, or None if safe.
-        pinned_ip: Resolved IP to connect to, preventing DNS rebinding
-            between validation and the actual HTTP request.
-    """
-
-    error: str | None
-    pinned_ip: str | None
-
-
-async def _validate_discovery_url(url: str) -> _SsrfCheckResult:
-    """Validate a URL for SSRF safety before making a discovery request.
-
-    Allows http/https schemes only and blocks private/reserved IP
-    addresses -- both literal IPs in the URL and resolved addresses
-    for hostnames (DNS rebinding protection).  Hostnames like
-    ``localhost`` are resolved via ``socket.getaddrinfo`` (offloaded
-    to a thread executor to avoid blocking the event loop) and checked
-    against the same blocked-network list.
-
-    On success, returns the resolved IP so the caller can pin the
-    connection to that address (preventing DNS rebinding between
-    validation and the actual HTTP request).
-
-    Args:
-        url: URL to validate.
-
-    Returns:
-        Check result with error message or pinned IP.
-    """
-    parsed = urlparse(url)
-
-    if parsed.scheme not in _ALLOWED_SCHEMES:
-        return _SsrfCheckResult(
-            f"scheme {parsed.scheme!r} not allowed; use http or https",
-            None,
-        )
-
-    hostname = parsed.hostname
-    if not hostname:
-        return _SsrfCheckResult("URL has no hostname", None)
-
-    return await _check_blocked_address(hostname)
-
-
-async def _check_blocked_address(hostname: str) -> _SsrfCheckResult:
-    """Check whether a hostname resolves to a blocked network range.
-
-    Handles both literal IPs and DNS names.  DNS resolution is
-    offloaded to a thread executor to avoid blocking the event loop.
-
-    Args:
-        hostname: Hostname or IP address string.
-
-    Returns:
-        Check result with error or the safe resolved IP.
-    """
-    # Fast path: literal IP address (no I/O).
-    try:
-        addr = ipaddress.ip_address(hostname)
-    except ValueError:
-        pass  # Not a literal IP -- resolve via DNS below.
-    else:
-        return _check_ip_blocked(addr, hostname)
-
-    # Resolve hostname and check all returned addresses.
-    return await asyncio.to_thread(_check_resolved_hostname, hostname)
-
-
-def _check_ip_blocked(
-    addr: ipaddress.IPv4Address | ipaddress.IPv6Address,
-    label: str,
-) -> _SsrfCheckResult:
-    """Check a single IP against blocked networks.
-
-    Args:
-        addr: IP address to check.
-        label: Display label for error messages.
-
-    Returns:
-        Check result with error or the safe IP string.
-    """
-    if isinstance(addr, ipaddress.IPv6Address) and addr.ipv4_mapped:
-        addr = addr.ipv4_mapped
-    for network in _BLOCKED_NETWORKS:
-        if addr in network:
-            return _SsrfCheckResult(
-                f"address {label!r} is in a blocked network range",
-                None,
-            )
-    return _SsrfCheckResult(None, str(addr))
-
-
-def _check_resolved_hostname(hostname: str) -> _SsrfCheckResult:
-    """Resolve a hostname and check all addresses against blocked networks.
-
-    Args:
-        hostname: DNS hostname to resolve.
-
-    Returns:
-        Check result with error or the first safe resolved IP.
-    """
-    try:
-        infos = socket.getaddrinfo(hostname, None)
-    except socket.gaierror:
-        return _SsrfCheckResult(
-            f"hostname {hostname!r} could not be resolved",
-            None,
-        )
-
-    for _, _, _, _, sockaddr in infos:
-        try:
-            addr = ipaddress.ip_address(sockaddr[0])
-        except ValueError:
-            continue
-        result = _check_ip_blocked(addr, hostname)
-        if result.error is not None:
-            return _SsrfCheckResult(
-                f"hostname {hostname!r} resolves to {sockaddr[0]!r} in a blocked range",
-                None,
-            )
-        # First safe address -- pin to it.
-        return _SsrfCheckResult(None, str(addr))
-
-    return _SsrfCheckResult(f"hostname {hostname!r} has no resolvable addresses", None)
 
 
 async def discover_models(
@@ -419,29 +271,6 @@ def _log_skip_counts(
         )
 
 
-def _build_pinned_url(
-    original_url: str,
-    pinned_ip: str,
-) -> tuple[str, str]:
-    """Build a URL with hostname replaced by a resolved IP.
-
-    Args:
-        original_url: Original URL with hostname.
-        pinned_ip: Resolved IP address to connect to.
-
-    Returns:
-        Tuple of (pinned_url, original_hostname) for Host header.
-    """
-    parsed = urlparse(original_url)
-    original_host = parsed.hostname or ""
-    port = parsed.port
-    # IPv6 literal must be bracketed in URLs.
-    ip_part = f"[{pinned_ip}]" if ":" in pinned_ip else pinned_ip
-    pinned_netloc = f"{ip_part}:{port}" if port else ip_part
-    pinned_url = urlunparse(parsed._replace(netloc=pinned_netloc))
-    return pinned_url, original_host
-
-
 async def _fetch_json_trusted(
     url: str,
     preset_name: str | None,
@@ -568,7 +397,7 @@ async def _validate_and_pin(
         Tuple of (pinned_url, original_host).  pinned_url is None
         if validation fails.
     """
-    check = await _validate_discovery_url(url)
+    check = await validate_discovery_url(url)
     if check.error is not None:
         logger.warning(
             PROVIDER_DISCOVERY_FAILED,
@@ -591,7 +420,7 @@ async def _validate_and_pin(
         )
         return None, ""
 
-    pinned_url, original_host = _build_pinned_url(url, pinned_ip)
+    pinned_url, original_host = build_pinned_url(url, pinned_ip)
     return pinned_url, original_host
