+++ tests/unit/api/test_startup_wiring.py
@@ -15,11 +15,9 @@ import structlog
 from structlog.typing import EventDict
 from typeguard import suppress_type_checks
 
+from synthorg.api._tunnel_wiring import resolve_tunnel_state_dir
 from synthorg.api.approval_store import ApprovalStore
-from synthorg.api.integrations_wiring import (
-    _resolve_tunnel_state_dir,
-    auto_wire_integrations,
-)
+from synthorg.api.integrations_wiring import auto_wire_integrations
 from synthorg.api.lifecycle import _wire_ontology_service
 from synthorg.api.lifecycle_builder import (
     _wire_approval_gate,
@@ -346,34 +344,34 @@ class TestTunnelUnconditionalWiring:
 
     def test_state_dir_env_resolves(self, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.setenv("SYNTHORG_TUNNEL_STATE_DIR", "/data/tunnel")
-        assert _resolve_tunnel_state_dir() == Path("/data/tunnel")
+        assert resolve_tunnel_state_dir() == Path("/data/tunnel")
 
     def test_state_dir_unset_yields_none(self, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.delenv("SYNTHORG_TUNNEL_STATE_DIR", raising=False)
-        assert _resolve_tunnel_state_dir() is None
+        assert resolve_tunnel_state_dir() is None
 
     def test_state_dir_traversal_rejected(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         monkeypatch.setenv("SYNTHORG_TUNNEL_STATE_DIR", "/data/../etc")
         with pytest.raises(ValueError, match="path traversal"):
-            _resolve_tunnel_state_dir()
+            resolve_tunnel_state_dir()
 
     def test_hostile_state_dir_degrades_wiring_to_none(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         """End-to-end: a traversal env value must not abort startup.
 
-        ``_wire_tunnel_provider`` is best-effort; the hostile value is
+        ``wire_tunnel_provider`` is best-effort; the hostile value is
         rejected inside the try block and the tunnel card degrades to
         unavailable rather than the whole boot failing.
         """
-        from synthorg.api.integrations_wiring import _wire_tunnel_provider
+        from synthorg.api._tunnel_wiring import wire_tunnel_provider
 
         monkeypatch.setenv("SYNTHORG_TUNNEL_STATE_DIR", "/data/../etc")
         config = RootConfig(company_name="test")
         with structlog.testing.capture_logs() as captured:
-            provider = _wire_tunnel_provider(config)
+            provider = wire_tunnel_provider(config)
         assert provider is None
         degrade_logs = [
             e
