+++ tests/unit/providers/test_discovery.py
@@ -7,11 +7,11 @@ from unittest.mock import AsyncMock, Mock, patch
 import httpx
 import pytest
 
-from synthorg.providers.discovery import (
-    _SsrfCheckResult,
-    _validate_discovery_url,
-    discover_models,
+from synthorg.providers._discovery_ssrf import (
+    SsrfCheckResult,
+    validate_discovery_url,
 )
+from synthorg.providers.discovery import discover_models
 from synthorg.providers.presets import LocalPreset
 from synthorg.providers.probing import (
     ProbeResult,
@@ -49,9 +49,9 @@ def _mock_client(
 @pytest.fixture(autouse=False)
 def _bypass_ssrf() -> Generator[None]:
     """Patch SSRF validation so HTTP-behavior tests can use localhost URLs."""
-    safe_result = _SsrfCheckResult(error=None, pinned_ip="127.0.0.1")
+    safe_result = SsrfCheckResult(error=None, pinned_ip="127.0.0.1")
     with patch(
-        "synthorg.providers.discovery._validate_discovery_url",
+        "synthorg.providers.discovery.validate_discovery_url",
         return_value=safe_result,
     ):
         yield
@@ -365,7 +365,7 @@ class TestValidateDiscoveryUrl:
     def _mock_dns(self) -> Generator[None]:
         """Provide deterministic DNS for URL validation tests."""
         with patch(
-            "synthorg.providers.discovery.socket.getaddrinfo",
+            "synthorg.providers._discovery_ssrf.socket.getaddrinfo",
             side_effect=_fake_getaddrinfo,
         ):
             yield
@@ -393,7 +393,7 @@ class TestValidateDiscoveryUrl:
         ],
     )
     async def test_url_validation(self, url: str, *, expected_safe: bool) -> None:
-        result = await _validate_discovery_url(url)
+        result = await validate_discovery_url(url)
         if expected_safe:
             assert result.error is None, (
                 f"Expected {url} to be safe, got: {result.error}"
@@ -701,7 +701,7 @@ class TestDiscoverModelsTrustedUrl:
                 "synthorg.providers.discovery.httpx.AsyncClient",
             ) as mock_cls,
             patch(
-                "synthorg.providers.discovery._validate_discovery_url",
+                "synthorg.providers.discovery.validate_discovery_url",
             ) as mock_ssrf,
         ):
             mock_cls.return_value = _mock_client(response)
